+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 ## [0.4.3-11] - 2020-01-20
+### Added
+- chapter and episode char count output
 ### Changed
 - scene setting shared over episodes
 ### Fixed