+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - chapter and episode char count output
 - block info output
+- stage line output
 ### Changed
 - scene setting shared over episodes
 ### Fixed
@@ -286,7 +287,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Example story as usage.
 - Output story as markdown.
 
-[Unreleased]: https://github.com/nagisc007/storybuilder/compare/v0.4.3-10...HEAD
+[Unreleased]: https://github.com/nagisc007/storybuilder/compare/v0.4.3-11...HEAD
+[0.4.3-11]: https://github.com/nagisc007/storybuilder/releases/v0.4.3-11
 [0.4.3-10]: https://github.com/nagisc007/storybuilder/releases/v0.4.3-10
 [0.4.3-9]: https://github.com/nagisc007/storybuilder/releases/v0.4.3-9
 [0.4.3-8]: https://github.com/nagisc007/storybuilder/releases/v0.4.3-8